requirement.txt
streamlit==1.28.0
numpy==1.24.3
Pillow==10.0.1
tensorflow==2.13.0
scikit-learn==1.3.0
h5py==3.9.0
pandas==2.1.1